requirements.txt
# Django y REST framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# MongoDB y MongoEngine
pymongo==4.5.0
mongoengine==0.27.0
django-mongoengine==0.5.6

# Celery y Redis
celery==5.3.4
redis==5.0.1

# OpenAI
openai==0.28.1

# Utilidades
python-dotenv==1.0.0
pyjwt==2.8.0

# Desarrollo y pruebas
pytest==7.4.3
pytest-django==4.7.0